Mod/#2: 테마 변경
- 필요 없는 색상 제거
- colorScheme 통일
- 외 package 이동
