+++ app/src/main/java/org/sopt/and/presentation/theme/Color.kt
@@ -1,14 +1,6 @@
-package org.sopt.and.ui.theme
+package org.sopt.and.presentation.theme
 
 import androidx.compose.ui.graphics.Color
-
-val Purple80 = Color(0xFFD0BCFF)
-val PurpleGrey80 = Color(0xFFCCC2DC)
-val Pink80 = Color(0xFFEFB8C8)
-
-val Purple40 = Color(0xFF6650a4)
-val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
 
 // Color palette
 val Background = Color(0xFF1B1B1B)
