+++ app/src/main/java/org/sopt/and/presentation/theme/Theme.kt
@@ -1,0 +1,45 @@
+package org.sopt.and.presentation.theme
+
+import android.app.Activity
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.lightColorScheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.SideEffect
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.toArgb
+import androidx.compose.ui.platform.LocalView
+import androidx.core.view.WindowCompat
+
+private val ColorScheme = lightColorScheme(
+    primary = WavveMain,
+    onPrimary = Color.White,
+    primaryContainer = Container,
+    onPrimaryContainer = Color.White,
+    surface = Background,
+    onSurface = LightGray,
+    background = Background,
+    onBackground = LightGray
+)
+
+@Composable
+fun ANDANDROIDTheme(
+    content: @Composable () -> Unit
+) {
+    val darkTheme: Boolean = isSystemInDarkTheme()
+    val colorScheme = ColorScheme
+    val view = LocalView.current
+    if (!view.isInEditMode) {
+        SideEffect {
+            val window = (view.context as Activity).window
+            window.statusBarColor = Color.Transparent.toArgb()
+            WindowCompat.getInsetsController(window, view).isAppearanceLightStatusBars = darkTheme
+        }
+    }
+
+    MaterialTheme(
+        colorScheme = colorScheme,
+        typography = Typography,
+        content = content
+    )
+}
